+++ CHANGELOG.md
@@ -15,6 +15,7 @@ We go to the next version after each release on
 
 ### Added
 
+- Dash app running visibility.
 - EGO donations.
 - Workspace peek (GNOME Shell 40 and higher).